+++ Runtime/Loaders/ResourcesUiAssetLoader.cs
@@ -40,7 +40,7 @@ namespace GameLovers.UiService
 		{
 			if (asset != null)
 			{
-				Object.Destroy(asset);
+				Object.DestroyImmediate(asset);
 			}
 		}
 	}